+++ src/ee/executors/ghash.h
@@ -53,22 +53,22 @@ void ghashWrapper(int block_x, int block_y,
 					);
 
 void indexCountWrapper(int block_x, int block_y,
-					int grid_x, int grid_y,
-					GNValue *outer_table,
-					int outer_rows,
-					int col_num,
-					uint64_t *searchPackedKey,
-					GTreeNode *searchKeyExp,
-					int *searchKeySize,
-					int searchExpNum,
-					uint64_t *packedKey,
-					uint64_t *bucketLocation,
-					uint64_t *hashedIndex,
-					int *indexCount,
-					int keySize,
-					int maxNumberOfBuckets,
-					GNValue *stack
-					);
+						int grid_x, int grid_y,
+						GNValue *outer_table,
+						int outer_rows,
+						int col_num,
+						uint64_t *searchPackedKey,
+						GTreeNode *searchKeyExp,
+						int *searchKeySize,
+						int searchExpNum,
+						uint64_t *packedKey,
+						uint64_t *bucketLocation,
+						uint64_t *hashedIndex,
+						ulong *indexCount,
+						int keySize,
+						int maxNumberOfBuckets,
+						GNValue *stack
+						);
 
 void hashJoinWrapper(int block_x, int block_y,
 						int grid_x, int grid_y,
